+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.1] - 2024-10-26
 - fix #3, setValue()
 - add **setPowerMode()** and **getPowerMode()**
+- fix #2, implement derived class **AD5640**
 - update unit test
 - update readme.md
 - minor edits.